Reorganize core Rust crates
