+++ nautilus_core/core/src/ffi/string.rs
